Improve cache manager

+++ plugins/nf-tower/src/main/io/seqera/tower/plugin/CacheManager.groovy
@@ -19,6 +19,7 @@ package io.seqera.tower.plugin
 
 import static java.nio.file.StandardCopyOption.*
 
+import java.nio.file.Files
 import java.nio.file.NoSuchFileException
 import java.nio.file.Path
 import java.nio.file.Paths
@@ -82,6 +83,11 @@ class CacheManager {
         if( !remoteWorkDir || !sessionUuid )
             return
 
+        if(!Files.exists(remoteCachePath)) {
+            log.debug "Remote cache path does not exist: $remoteCachePath - skipping cache restore"
+            return
+        }
+
         try {
             log.info "Restoring cache: ${remoteCachePath.toUriString()} => ${localCachePath.toUriString()}"
             localCachePath.deleteDir()
@@ -89,13 +95,18 @@ class CacheManager {
             FileHelper.copyPath(remoteCachePath, localCachePath, REPLACE_EXISTING)
         }
         catch (NoSuchFileException e) {
-            log.info "Remote cache restore ignored -- reason: ${e.message ?: e}"
+            log.info "Remote cache restore ignored — reason: ${e.message ?: e}"
         }
     }
 
     protected void saveCacheFiles() {
         if( !remoteWorkDir || !sessionUuid )
             return
+
+        if( !Files.exists(localCachePath) ) {
+            log.debug "Local cache path does not exist: $localCachePath — skipping cache backup"
+            return
+        }
 
         // upload nextflow cache metadata
         try {
@@ -105,47 +116,48 @@ class CacheManager {
             FilesEx.copyTo(localCachePath, remoteCachePath)
         }
         catch (Throwable e) {
-            log.warn "Failed to backup resume metadata to remote store path: ${remoteCachePath.toUriString()}", e
+            log.warn "Failed to backup resume metadata to remote store path: ${remoteCachePath.toUriString()} — cause: ${e}", e
         }
-        // -- upload out file
+
+        // — upload out file
         try {
             if( localOutFile?.exists() )
                 FileHelper.copyPath(localOutFile, remoteOutFile, REPLACE_EXISTING)
         }
         catch (Throwable e) {
-            log.warn "Unable to upload nextflow out file: $localOutFile -- reason: ${e.message ?: e}", e
+            log.warn "Unable to upload nextflow out file: $localOutFile — reason: ${e.message ?: e}", e
         }
-        // -- upload log file
+        // — upload log file
         try {
             if( localLogFile?.exists() )
                 FileHelper.copyPath(localLogFile, remoteLogFile, REPLACE_EXISTING)
         }
         catch (Throwable e) {
-            log.warn "Unable to upload nextflow log file: $localLogFile -- reason: ${e.message ?: e}", e
+            log.warn "Unable to upload nextflow log file: $localLogFile — reason: ${e.message ?: e}", e
         }
-        // -- upload timeline file
+        // — upload timeline file
         try {
             if( localTimelineFile?.exists() )
                 FileHelper.copyPath(localTimelineFile, remoteTimelineFile, REPLACE_EXISTING)
         }
         catch (Throwable e) {
-            log.warn "Unable to upload nextflow timeline file: $localTimelineFile -- reason: ${e.message ?: e}", e
+            log.warn "Unable to upload nextflow timeline file: $localTimelineFile — reason: ${e.message ?: e}", e
         }
-        // -- upload tower config file
+        // — upload tower config file
         try {
             if( localTowerConfig?.exists() )
                 FileHelper.copyPath(localTowerConfig, remoteTowerConfig, REPLACE_EXISTING)
         }
         catch (Throwable e) {
-            log.warn "Unable to upload tower config file: $localTowerConfig -- reason: ${e.message ?: e}", e
+            log.warn "Unable to upload tower config file: $localTowerConfig — reason: ${e.message ?: e}", e
         }
-        // -- upload tower reports file
+        // — upload tower reports file
         try {
             if( localTowerReports?.exists() )
                 FileHelper.copyPath(localTowerReports, remoteTowerReports, REPLACE_EXISTING)
         }
         catch (Throwable e) {
-            log.warn "Unable to upload tower reprts file: $localTowerReports -- reason: ${e.message ?: e}", e
+            log.warn "Unable to upload tower reprts file: $localTowerReports — reason: ${e.message ?: e}", e
         }
     }
